troff/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
SQLAlchemy==2.0.43
apscheduler==3.11.0
requests==2.32.5
python-slugify==8.0.4
python-multipart==0.0.20
openai==1.107.1
beautifulsoup4==4.12.3
google-cloud-storage==2.18.2
Pillow==11.0.0
